refactor: changed the name of the app icon component folder and the imports
