+++ src/providers/index.tsx
@@ -5,9 +5,6 @@ import { ReactNode, useState } from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import Step1Provider from '@/context/Step1Context';
-import Step2Provider from '@/context/Step2Context';
-import Step3Provider from '@/context/Step3Context';
 import UserIdProvider from '@/context/UserIdContext';
 import { queryClientConfig } from '@/lib/queryClient';
 
@@ -16,14 +13,8 @@ const Providers = ({ children }: { children: ReactNode }) => {
   return (
     <QueryClientProvider client={queryClient}>
       <UserIdProvider>
-        <Step1Provider>
-          <Step2Provider>
-            <Step3Provider>
-              {children}
-              <ReactQueryDevtools initialIsOpen={false} />
-            </Step3Provider>
-          </Step2Provider>
-        </Step1Provider>
+        {children}
+        <ReactQueryDevtools initialIsOpen={false} />
       </UserIdProvider>
     </QueryClientProvider>
   );
