+++ src/repositories/subscriptions/index.ts
@@ -1,4 +1,3 @@
-import { SubscriptionPayload } from '@/components/parts/SubscriptionForms/type';
 import { fetcher } from '@/lib/fetcher';
 
 export const getAllSubscription = async () => {
@@ -16,6 +15,16 @@ export const getSubscriptionById = async (id: string) => {
   return response;
 };
 
+export const addSubscription = async (payload: Record<string, unknown>) => {
+  const response = await fetcher({
+    url: '/subscriptions',
+    method: 'POST',
+    body: JSON.stringify(payload)
+  });
+
+  return response;
+};
+
 export const editSubscription = async (id: string, payload: Record<string, unknown>) => {
   const response = await fetcher({
     url: `/subscriptions/${id}`,
@@ -30,16 +39,6 @@ export const deleteSubscription = async (id: string) => {
   const response = await fetcher({
     url: `/subscriptions/${id}`,
     method: 'DELETE'
-  });
-
-  return response;
-};
-
-export const addSubscription = async (payload: SubscriptionPayload) => {
-  const response = await fetcher({
-    url: '/subscriptions',
-    method: 'POST',
-    body: JSON.stringify(payload)
   });
 
   return response;
